Добавлено удаление старых дизлайков для коливингов
